backend/requirements-deploy.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pdfplumber>=0.11.0
python-docx>=1.1.0
groq>=0.9.0
httpx>=0.27.0
aiofiles>=23.2.1
aiohttp>=3.9.5
livekit-api>=0.7.0
Pillow>=10.0.0
